Final version

See merge request hpc-ai_21/cpp-life-of-boids!53

+++ src/lib/myLaws/Law.cpp
@@ -11,7 +11,7 @@ std::vector<Vec2> Law::getCoordinatesArray(const std::vector<Agent>& neighbors) 
 	//We create a new array with a size of the number of neighbors
 	std::vector<Vec2> points(neighbors.size());
 
-	for (int i = 0; i < neighbors.size(); ++i) {
+	for (int i = 0; i < (int)neighbors.size(); ++i) {
 		//filling the array with the coordinates of the position of each agent of the neighborhood
 		float x = neighbors[i].getPosition().x;
 		float y = neighbors[i].getPosition().y;
